shootDelay update, + RPM system
Updated all shooting delay variables to floats.
Added "RoundsPerMin" as new config setting. Converts RPM to tick delay float offering more flexibility. (Replaces ShootDelay)

+++ src/main/java/com/flansmod/client/FlansModClient.java
@@ -84,7 +84,7 @@ public class FlansModClient extends FlansMod
 	public static int controlModeSwitchTimer = 20;
 	
 	/** The delay between shots / reloading */
-	public static int shootTimeLeft, shootTimeRight;
+	public static float shootTimeLeft, shootTimeRight;
 	
 	//Recoil variables
 	/** The recoil applied to the player view by shooting */
@@ -256,7 +256,7 @@ public class FlansModClient extends FlansMod
 
 	}
 	
-	public static int shootTime(boolean left)
+	public static float shootTime(boolean left)
 	{
 		return left ? shootTimeLeft : shootTimeRight;
 	}
